input_sniffer: Drop "bytes" key
Since 9e95407 (new dnstap output, 2020-12-30) the sniffer input has had
the payload in the "payload" key to match the dnstap output, meaning
that the "bytes" key is just duplicate data now.

This fixes a bug where the sniffer input cannot be used with either a
JSON or YAML format output, since the transform function doesn't remove
the "bytes" key.

Signed-off-by: Johannes Löthberg <johannes.lothberg@tele2.com>